% Generated by roxygen2 (4.0.0): do not edit by hand
\name{calcbeta_stat}
\alias{calcbeta_stat}
\title{Calculate beta-diversity statistic for a single realization}
\usage{
calcbeta_stat(m, ttt, method = "jaccard", binary = TRUE, stat = c("Fstat",
  "Fstat_perm", "pval_perm", "pval_Fstat_perm"))
}
\arguments{
\item{m}{community matrix}

\item{ttt}{treatment/grouping factor}

\item{method}{beta diversity metric}

\item{binary}{compute beta diversity based on presence/absence?}

\item{stat}{beta diversity statistic to store: \code{Fstat}=F-statistic from \code{\link{anova.betadisper}}; \code{Fstat_perm}=F-statistic from \code{\link{permutest.betadisper}}; \code{pval_perm}=p-value from \code{\link{permutest.betadisper}}; \code{pval_Fstat_perm}=F- and p-values as above}
}
\description{
Calculate beta-diversity statistic for a single realization
}

