% Generated by roxygen2 (4.0.0): do not edit by hand
\name{findTargets}
\alias{findTargets}
\title{Find target sizes for rarefaction of samples for multiple treatments
down to a common size.}
\usage{
findTargets(nList, method = c("rankMatch", "randomMatch", "smallest"),
  maxit = 1000, debug = FALSE)
}
\arguments{
\item{nList}{a list of numeric vectors of patch sizes (total abundances) for each treatment}

\item{method}{(character): 'smallest' to use the smallest abundance found in any patch in any treatment; 'rankMatch' to match patches by rank within treatment and use the smallest abundance within a rank category; 'randomMatch' to match patches into groups at random. The idea of the default 'rankMatch'
is that by matching patches by rank we rarefy as
little as necessary to get equal patch sizes in the two
groups -- i.e. the minimum in both groups is the same, etc.
(I can't prove that this minimizes the required
amount of rarefaction, and I'm not even sure it's always
true, but in any case it seems to be a reasonable approach.)}

\item{maxit}{number of iterations to try for randomized algorithms}

\item{debug}{debugging output?}
}
\value{
a list of vectors of target sizes corresponding to the patches described in \code{nList}
}
\description{
Find target sizes for rarefaction of samples for multiple treatments
down to a common size.
}
\examples{
## make sure that the order of the patches (i.e. order of number
## of fish in each patch) doesn't matter
nList <- list(1:10,10:1)
findTargets(nList)  ## unchanged
## three treatments
nList <- list(5:10,9:4,c(2,7,8,1,20,8))
f1 <- findTargets(nList)
## unequal numbers of patches
nList <- list(5:10,1:10)
set.seed(101)
(f2 <- findTargets(nList))
}

