% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.spmat}
\alias{plot.commframe}
\alias{plot.spmat}
\title{Plot species matrices and community frames}
\usage{
\method{plot}{spmat}(x, y, aspect = "fill", ...)

\method{plot}{commframe}(x, y, aspect = "fill", ...)
}
\arguments{
\item{x}{a species matrix (class \code{spmat}) or community data frame (class \code{commframe})}

\item{y}{ignored, for generic compatibility}

\item{\dots}{ignored, for generic compatibility}

\item{aspect}{aspect ratio (\code{"fill"} or \code{"iso"})}
}
\description{
Plot species matrices and community frames
}

