% Generated by roxygen2 (4.0.0): do not edit by hand
\name{sampleMat}
\alias{sampleMat}
\title{Resample a community matrix}
\usage{
sampleMat(x, n.target)
}
\arguments{
\item{x}{a species abundance matrix to rarefy}

\item{n.target}{rarefied size for each sample}
}
\value{
a matrix of the same structure as the original
}
\description{
Resample a community matrix
}

