% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rarefaction_funs.R
\name{simComm}
\alias{simComm}
\title{simulate multiple communities with different beta/abundance/etc.
characteristics}
\usage{
simComm(n.indiv.site = rep(10, ncomm), n.site = rep(3, ncomm),
  spcat = rep(1, ncomm), ncomm = 2, p.mix = rep(0.5, ncomm),
  n.abund = rep(1, ncomm), totsp = rep(3, ncomm), rand = "poisson",
  commlabs = letters[1:ncomm], retval = c("combined", "list"),
  seed = NULL)
}
\arguments{
\item{n.indiv.site}{individuals per site}

\item{n.site}{number of sites}

\item{spcat}{number of species per abundance category PER SITE (overdetermined)}

\item{ncomm}{number of communities}

\item{p.mix}{mixing parameter: probability of among-site mixing for each abundance category: 0=endemic, 1=homogeneous}

\item{n.abund}{number of abundance categories}

\item{totsp}{total species pool size (FIXME)}

\item{rand}{randomization type: 'none' to keep numbers==expected numbers; 'multinom' for fixed total per site; 'poisson' for Poisson sampling, 'trpoisson' for truncated Poisson sampling}

\item{commlabs}{labels for communities/treatments}

\item{retval}{return communities as a single metacommunity matrix or as a list of community matrices?}

\item{seed}{random-number seed}
}
\description{
simulate multiple communities with different beta/abundance/etc.
characteristics
}
\examples{
plot(simComm(totsp=c(30,30),p.mix=c(0.2,0.7)))
}

