% Generated by roxygen2 (4.0.1): do not edit by hand
\name{simfun0}
\alias{simfun0}
\alias{simfun1}
\alias{simfun2}
\title{rarefy once (maybe) and spit out F statistic(s)}
\usage{
simfun0(matList, tList, ttt, rarefy = TRUE, bstat = "Fstat")

simfun1(matList, tList, ttt, bstat, rarefy = TRUE, .progress = "text",
  nsim = 100)

simfun2(..., bstat, rarefy = TRUE, .progress = "text", nsim = 100)
}
\arguments{
\item{matList}{list of community matrices}

\item{tList}{list of target sizes for each community}

\item{ttt}{treatment vector}

\item{bstat}{beta diversity statistic (see \code{\link{calcbeta_stat}})}

\item{rarefy}{(logical) perform rarefaction?}

\item{.progress}{passed to \code{simfun1}}

\item{nsim}{number of simulations}

\item{\dots}{arguments passed to \code{simfun1}}
}
\description{
rarefy once (maybe) and spit out F statistic(s)

return vector (or matrix) of F statistics

simulate one community, then
rarefy multiple times and return the vector (or matrix) of F statistics
}

