% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betasim_funs.R
\name{betasim}
\alias{betasim}
\title{Simulate communities with specified diversity characteristics}
\usage{
betasim(n.abund = 5, p.abund = 0.5, diff.abund = NULL, spcat = 2,
  totsp = 30, n.site = 3, n.indiv.site = 100, n.indiv.tot = NULL,
  p.mix = rep(0, n.abund), seed = NULL, rand = c("none", "multinom",
  "poisson", "trpoisson"), rarefy = 1, pred.spec = c("none", "focal",
  "abund"), pred.rand = c("binom", "none"), pred.gamma = -Inf,
  pred.alpha = 0, pred.gamma.sd = 0, no.zero.patch = TRUE)
}
\arguments{
\item{n.abund}{number of abundance categories}

\item{p.abund}{relative ranking of subsequent abundance categories}

\item{diff.abund}{alternative parameterization: difference between most and least abundant category}

\item{spcat}{number of species per abundance category PER SITE (overdetermined)}

\item{totsp}{total species pool size (FIXME)}

\item{n.site}{number of sites}

\item{n.indiv.site}{individuals per site}

\item{n.indiv.tot}{total number of individuals}

\item{p.mix}{mixing parameter: probability of among-site mixing for each abundance category: 0=endemic, 1=homogeneous}

\item{seed}{random number seed}

\item{rand}{randomization type: 'none' to keep numbers==expected numbers; 'multinom' for fixed total per site; 'poisson' for Poisson sampling, 'trpoisson' for truncated Poisson sampling}

\item{rarefy}{degree of rarefaction (1=none)}

\item{pred.spec}{predator specification (stub?)}

\item{pred.rand}{predator randomness}

\item{pred.gamma}{predation intensity (-Inf==none)}

\item{pred.alpha}{predator preference parameter; if \code{pred.spec=="abund"} 0=neutral, -1=pref. rare, +1=pref. common; if pred.spec=="focal" 0=neutral, -1=pref non-endemic, +1=pref endemic}

\item{pred.gamma.sd}{patch-to-patch variability in predation intensity (0=none)}

\item{no.zero.patch}{force all patches to have positive occupancy?}
}
\description{
Simulate communities with specified diversity characteristics
}

