% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calcbeta}
\alias{calcbeta}
\title{compute beta diversity summary statistic}
\usage{
calcbeta(m, method = "jaccard", distances = c("centroid", "pairwise"),
  trap.errors = TRUE, transFun = identity, binaryFlag = FALSE, ...)
}
\arguments{
\item{m}{community matrix}

\item{method}{diversity measure}

\item{distances}{calculate average distance to centroid or average pairwise distance?}

\item{trap.errors}{replace error returns with NA values?}

\item{transFun}{transform (e.g. for taking square-root of abundances)}

\item{binaryFlag}{compute presence.absence statistics?}

\item{\dots}{additional parameters for \code{raupcrick}}
}
\description{
compute beta diversity summary statistic
}

