% Generated by roxygen2 (4.0.0): do not edit by hand
\name{raref_test}
\alias{raref_test}
\alias{raref_test2}
\title{rarefaction test}
\usage{
raref_test(comm, ttt, .progress = "text", n_raref = 50, type = c("Fmean",
  "pmean", "pbinom", "singlecomp"), rarefy = TRUE)

raref_test2(comm, ttt, n_raref = 50, nperm = 200, method = "jaccard",
  binary = TRUE, type = c("raref_perm_meanF", "raref_perm_meanp",
  "raref_perm_medianp"), return.all = FALSE, dummy_raref = FALSE)
}
\arguments{
\item{comm}{community matrix (patches=rows, species=columns)}

\item{ttt}{treatment vector (treatment id of each patch)}

\item{.progress}{progress bar type -- passed through to \code{\link{raply}}}

\item{n_raref}{number of rarefactions}

\item{type}{method}

\item{rarefy}{actually perform rarefaction?}

\item{dummy_raref}{fake rarefaction?}

\item{return.all}{return all information?}

\item{nperm}{number of permutations}

\item{method}{beta diversity metric}

\item{binary}{use presence/absence?}
}
\description{
rarefaction test
}

