% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rarefaction_funs.R
\name{sampleMat}
\alias{sampleMat}
\title{Individual-based resampling/rarefaction of a community matrix}
\usage{
sampleMat(x, n.target, debug = FALSE)
}
\arguments{
\item{x}{a species abundance matrix to rarefy (rows=sites, columns=species)}

\item{n.target}{rarefied size for each sample}

\item{debug}{(logical) debugging flag}
}
\value{
a matrix of the same structure as the original
}
\description{
Individual-based resampling/rarefaction of a community matrix
}

